fix(react-native-auth): add TurboModule spec
