+++ .changeset/purple-dots-travel.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/react-native-auth": patch
+---
+
+Add TurboModule spec
